Inheritance simplification: Generalized Entity and added a class in between Archer and Entity (Character)
Entity is now simply something with a sprite that can move. How it moves
must be defined by the child class (e.g. uses path finding or not).

Character is a type of Entity that has health and must path find through
the terrain.

Archer is a type of Character with a specific behavior. More refinement
needed.

Arrow will eventually become a type of Entity that moves without path
finding. More design required.